requirements.txt
anthropic
python-dotenv
watchdog
google-genai
pydantic
